pyproject.toml
[tool.poetry]
name = "Cthulhu"
version = "0.0.1"
description = "De todito"
authors = ["Mario Hernandez <yo@mariofix.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1"
thankyou = "^0.0.1"
PyMySQL = "^1.0.2"
django-baton = "^2.4.1"
django-debug-toolbar = "^3.6.0"
django-logentry-admin = "^1.1.0"
django-registration = "^3.3"
celery = "^5.2.7"
django-celery-results = "^2.4.0"
redis = "^4.3.4"
python-dotenv = "^0.20.0"
bcrypt = "^3.2.2"
requests = "^2.28.1"
jsonfield = "^3.1.0"
sentry-sdk = "^1.10.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "*"
pytest = "*"
pytest-cov = "*"

[tool.poetry.group.deploy]
optional = true

[tool.poetry.group.deploy.dependencies]
gunicorn = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
